teststreamlit/requirements.txt
import streamlit
pandas
numpy
matplotlib.pyplot
plotly
numpy
